revisited test_absorbing_boundary_validation.py
- renamed folders test_lysmer_boundary_column3d_tetra, test_lysmer_boundary_column3d_tetra_hor, test_lysmer_column3d_hexa and placed in lysmer_boundary_column3D
- renamed test_absorbing_boundary_validation.py into absorbing_boundary_3D.py
