--- applications/GeoMechanicsApplication/tests/test_lysmer_boundary_column3d_tetra_hor.gid/MaterialParameters.json
@@ -1,36 +1,0 @@
-{
-   "properties": [{
-      "model_part_name":         "PorousDomain.soil",
-      "properties_id":           1,
-      "Material": {
-          "constitutive_law": {
-              "name"             :  "LinearElastic3DLaw" 
-          },
-          "Variables": {
-              "IGNORE_UNDRAINED"         :  true,
-              "YOUNG_MODULUS"            :  10000,
-              "POISSON_RATIO"            :  0.2,
-              "DENSITY_SOLID"            :  2.65,
-              "DENSITY_WATER"            :  1.0,
-              "POROSITY"                 :  0.3,
-              "BULK_MODULUS_SOLID"       :  1.0e9,
-              "BULK_MODULUS_FLUID"       :  2.0e-30,
-              "PERMEABILITY_XX"          :  4.5e-30,
-              "PERMEABILITY_YY"          :  4.5e-30,
-              "PERMEABILITY_ZZ"          :  4.5e-30,
-              "PERMEABILITY_XY"          :  0.0,
-              "PERMEABILITY_YZ"          :  0.0,
-              "PERMEABILITY_ZX"          :  0.0,
-              "DYNAMIC_VISCOSITY"        :  8.90e-7,
-              "RETENTION_LAW"                    : "SaturatedBelowPhreaticLevelLaw",
-              "SATURATED_SATURATION"             :  1.0,
-              "RESIDUAL_SATURATION"              :  1e-10,
-              "VAN_GENUCHTEN_AIR_ENTRY_PRESSURE" :  2.561,
-              "VAN_GENUCHTEN_GN"                 :  1.377,
-              "VAN_GENUCHTEN_GL"                 :  1.25,
-              "MINIMUM_RELATIVE_PERMEABILITY"    :  0.0001
-         },
-         "Tables": {}
-      }
-   }]
-}
